Tool_compare_Batiment/requirements_full.txt
streamlit==1.28.2
PyMuPDF==1.23.9
openai==1.12.0
pillow==10.2.0
python-dotenv==1.0.0
streamlit-pdf-viewer==0.0.15
